+++ tests/test_latex.py
@@ -1,8 +1,10 @@
 # 3rd party
+from sphinx import addnodes
 from sphinx.events import EventListener
 
 # this package
 from sphinx_toolbox import latex
+from sphinx_toolbox.latex import succinct_seealso
 from sphinx_toolbox.testing import run_setup
 
 
@@ -21,6 +23,25 @@ def test_setup():
 
 	assert app.registry.source_parsers == {}
 	assert app.registry.translation_handlers["latex"]["footnote"] == (latex.visit_footnote, latex.depart_footnote)
+
+	assert app.registry.css_files == []
+	assert app.registry.js_files == []
+
+
+def test_setup_succinct_seealso():
+	setup_ret, directives, roles, additional_nodes, app = run_setup(succinct_seealso.setup)
+
+	assert additional_nodes == {addnodes.seealso}
+
+	assert app.events.listeners == {}
+
+	assert directives == {}
+
+	assert app.registry.source_parsers == {}
+	assert app.registry.translation_handlers["latex"]["seealso"] == (
+			succinct_seealso.visit_seealso,
+			succinct_seealso.depart_seealso,
+			)
 
 	assert app.registry.css_files == []
 	assert app.registry.js_files == []